Move lazyFetchEmailWithTimeout into own lib
